Re-organize raceinfo
- Move resolver out of raceinfo module
- Combine Time, Heat, Lane in single file
