+++ raceinfo/__init__.py
@@ -43,28 +43,24 @@ data file types:
 
 from .colorado_scb import load_all_scb, parse_scb, parse_scb_file
 from .dolphin_do4 import parse_do4, parse_do4_file
-from .heat import Heat
-from .lane import Lane
 from .nameformat import NameMode, format_name
-from .resolver import standard_resolver
 from .startlist import StartList, startlists_to_csv
-from .time import MIN_VALID_TIME, ZERO_TIME, Time, truncate_hundredths
+from .time import ZERO_TIME, Heat, Lane, Time, TimeResolver, truncate_hundredths
 
 __all__ = [
-    "MIN_VALID_TIME",
     "ZERO_TIME",
     "Heat",
     "Lane",
     "NameMode",
     "StartList",
     "Time",
+    "TimeResolver",
     "format_name",
     "load_all_scb",
     "parse_do4",
     "parse_do4_file",
     "parse_scb",
     "parse_scb_file",
-    "standard_resolver",
     "startlists_to_csv",
     "truncate_hundredths",
 ]
