+++ raceinfo/time.py
@@ -16,7 +16,12 @@
 
 """Fundamental timing types."""
 
+import copy
 import decimal
+import re
+from dataclasses import InitVar, dataclass, field
+from datetime import datetime
+from typing import Callable
 
 # https://docs.python.org/3/library/decimal.html#signals
 # Raise decimal.FloatOperation when Decimal and float are improperly mixed
@@ -26,8 +31,384 @@ Time = decimal.Decimal
 """A Time in seconds and hundredths of a second"""
 
 ZERO_TIME = Time("0.00")
-MIN_VALID_TIME = Time("10.00")
-"""The minimum valid race or split time."""
+
+type TimeResolver = Callable[[Lane], None]
+"""
+A function that resolves times for a `Lane`.
+
+The resolver should use a combination of the `primary`, `backups`, and `splits`
+attributes to determine a final time for the lane. That final time should be set in
+the `final_time` attribute of the `Lane`.
+"""
+
+
+@dataclass(kw_only=True)
+class Lane:
+    """
+    The per-lane information for a heat.
+
+    All fields are optional. If a field is not present, it is assumed to not be
+    supported by the timing system that generated this lane data. For example,
+    an empty ("") `team` name indicates that the system supports team names, but
+    the it was not provided for this lane. This is different than a `None`
+    value, which indicates that the system does not support team names at all.
+    The same logic applies to `backups` and `splits`. An empty `list` is
+    different than `None`.
+
+    :param name: The name of the swimmer
+    :param team: The name of the swimmer's team
+    :param seed_time: The swimmer's seed time
+    :param age: The swimmer's age
+    :param primary: The primary (pad) time for this lane
+    :param backups: The backup times for this lane
+    :param splits: The split times for this lane
+    :param is_dq: True if the swimmer was disqualified
+    :param is_empty: True if the lane is marked empty
+    :raises: ValueError if any of the parameters are invalid
+    """
+
+    # Swimmer information
+    name: str | None = None
+    """The name of the swimmer"""
+    team: str | None = None
+    """The name of the swimmer's team"""
+    seed_time: Time | None = None
+    """The swimmer's seed time"""
+    age: int | None = None
+    """The swimmer's age"""
+
+    # Race data
+    primary: Time | None = None
+    """The primary (pad) time"""
+    backups: list[Time | None] | None = None
+    """
+    The list of backup times.
+
+    - If backups is None, the timing system does not support backup times.
+    - If an entry in the list is None, the timing system did not capture a
+      backup time for that particular slot.
+    """
+    splits: list[Time | None] | None = None
+    """The list of intermediate, cumulative split times"""
+    final_time: Time | None = None
+    """
+    The resolved final time for this lane, if available
+
+    This value is not set by the timing system, but is generated by a resolver
+    that integrates the primary, backup, and potentially split times.
+    """
+
+    # Flags
+    is_dq: bool | None = None
+    """True if the swimmer was disqualified"""
+    is_empty: bool | None = None
+    """True if the lane is marked empty"""
+
+    def __post_init__(self):
+        """Validate the lane data."""
+        if self.age is not None and self.age < 0:
+            raise ValueError("Age must be non-negative")
+        if self.seed_time is not None and self.seed_time < ZERO_TIME:
+            raise ValueError("Seed time must be non-negative")
+        if self.primary is not None and self.primary < ZERO_TIME:
+            raise ValueError("Primary time must be non-negative")
+        if self.final_time is not None and self.final_time < ZERO_TIME:
+            raise ValueError("Final time must be non-negative")
+        for backup in self.backups or []:
+            if backup is not None and backup < ZERO_TIME:
+                raise ValueError("Backup times must be non-negative")
+        for split in self.splits or []:
+            if split is not None and split < ZERO_TIME:
+                raise ValueError("Splits must be non-negative")
+
+    @property
+    def is_noshow(self) -> bool:
+        """
+        Check if the lane is a no-show.
+
+        A lane is considered a no-show if it doesn't have any valid primary,
+        backup, or split times.
+        """
+        return (
+            (self.primary is None)
+            and (self.backups is None or all(backup is None for backup in self.backups))
+            and (self.splits is None or all(split is None for split in self.splits))
+        )
+
+    def merge(
+        self,
+        info_from: "Lane | None" = None,
+        results_from: "Lane | None" = None,
+    ) -> None:
+        """
+        Merge another Lane object into this one.
+
+        Merging "info" will overwrite:
+        - name, team, seed_time, age
+
+        Merging "results" will overwrite:
+        - primary, backups, splits, is_dq, is_empty
+
+        :param info_from: Merge the heat information into this one
+        :param results_from: Merge the race data into this one
+        """
+        if info_from is not None:
+            self.name = info_from.name
+            self.team = info_from.team
+            self.seed_time = info_from.seed_time
+            self.age = info_from.age
+
+        if results_from is not None:
+            self.primary = results_from.primary
+            self.backups = copy.deepcopy(results_from.backups)
+            self.splits = copy.deepcopy(results_from.splits)
+            self.final_time = results_from.final_time
+            self.is_dq = results_from.is_dq
+            self.is_empty = results_from.is_empty
+
+    def is_similar_to(self, other: "Lane") -> bool:
+        """
+        Check if this lane is similar to another lane.
+
+        Similarity is defined as having the same values for all fields where a
+        field is not None. If a field is None, it is ignored in the comparison.
+
+        Note: This method is mainly intended for testing purposes.
+
+        :param other: The other lane to compare to
+        :returns: True if the lanes are similar, False otherwise
+        """
+        # Compare the simple fields
+        for var in [
+            "name",
+            "team",
+            "seed_time",
+            "age",
+            "primary",
+            "final_time",
+            "is_dq",
+            "is_empty",
+        ]:
+            if (
+                getattr(self, var) is not None
+                and getattr(other, var) is not None
+                and getattr(other, var) != getattr(self, var)
+            ):
+                return False
+        # Compare the splits
+        if self.splits is not None and other.splits is not None:
+            if len(self.splits) != len(other.splits):
+                return False
+            for split1, split2 in zip(self.splits, other.splits):
+                if split1 != split2:
+                    return False
+        # Compare the backups
+        if self.backups is not None and other.backups is not None:
+            if len(self.backups) != len(other.backups):
+                return False
+            for backup1, backup2 in zip(self.backups, other.backups):
+                if backup1 != backup2:
+                    return False
+        return True
+
+
+@dataclass(kw_only=True)
+class Heat:
+    """
+    Information describing a heat.
+
+    Heat can be used for both heat information (e.g. from a start list) and for
+    race results (e.g. times from a timing system). Each supported data source
+    (start list file or timing system) will store its data in a Heat object. The
+    merge() method can be used to combine the data from both sources into a
+    single Heat object.
+
+    All fields are optional. If a field is `None`, it is assumed to not be
+    supported by the data source that generated this heat data.
+
+    :param event: The event number
+    :param heat: The heat number
+    :param description: The name of the event
+    :param meet_id: The meet ID
+    :param race: The race number
+    :param time_recorded: The time the results were recorded
+    :param time_resolver: The function to resolve times for this heat
+    :param lanes: The data for each lane
+    :raises: ValueError if any of the parameters are invalid
+    """
+
+    # Heat information
+    event: str | None = None
+    """
+    The an alphanumeric event number
+
+    This is a string to handle events like '1S' or '1Z' in the case of swim-offs
+    or initial splits.
+    """
+    heat: int | None = None
+    """The heat number"""
+    description: str | None = None
+    """The name of the event (e.g. 'Girls 100 Freestyle')"""
+
+    # Race data
+    meet_id: str | None = None
+    """The meet ID"""
+    race: int | None = None
+    """The race number"""
+    time_recorded: datetime | None = None
+    """The time the results were recorded"""
+
+    # We hide the actual lane data and provide lane() to access it to avoid
+    # confusion over the indexing of the array vs. the actual lane number.
+    lanes: InitVar[list[Lane] | None] = field(default=None)
+    """The data for each lane"""
+
+    def __post_init__(self, lanes: list[Lane] | None):
+        """Validate the heat data."""
+        self._lanes = lanes if lanes is not None else []
+        # Ensure that we have 10 lanes, even if they are empty
+        if len(self._lanes) > 10:  # noqa: PLR2004
+            raise ValueError("The maximum number of lanes is 10")
+        while len(self._lanes) < 10:  # noqa: PLR2004
+            self._lanes.append(Lane())
+        if self.event is not None:
+            self.event = self.event.upper()
+        if self.heat is not None and self.heat < 1:
+            raise ValueError("Heat number must be positive")
+        if self.race is not None and self.race < 1:
+            raise ValueError("The race number must be positive")
+
+    def lane(self, lane_number: int) -> Lane:
+        """
+        Retrieve the lane object for a given lane number.
+
+        :param lane_number: The lane number (1-10)
+        :returns: The lane object
+        :raises: ValueError if the lane number is invalid
+        """
+        if lane_number < 1 or lane_number > 10:  # noqa: PLR2004
+            raise ValueError("Lane number must be between 1 and 10")
+        return self._lanes[lane_number - 1]
+
+    def place(self, lane_number: int) -> int | None:
+        """
+        Retrieve the place for a given lane number.
+
+        The place is determined by the resolved time of the lane. If the lane is
+        empty, disqualified, or has an indeterminate final time, None is
+        returned.
+
+        **Note:** Times must be resolved before calling this method.
+
+        :param lane_number: The lane number (1-10)
+        :returns: The place for the lane or None
+        :raises: ValueError if the lane number is invalid
+        """
+        lane = self.lane(lane_number)
+        if lane.is_dq or lane.is_empty or lane.final_time is None:
+            return None
+        place = 1
+        for other_lane in self._lanes:
+            other_time = other_lane.final_time
+            if other_lane.is_empty or other_lane.is_dq or other_time is None:
+                continue
+            if other_time < lane.final_time:
+                place += 1
+        return place
+
+    def resolve_times(self, resolver: TimeResolver) -> None:
+        """
+        Resolve the times for all lanes in this heat.
+
+        :param resolver: The function to resolve times for this heat
+        """
+        for lane in self._lanes:
+            resolver(lane)
+
+    def has_names(self) -> bool:
+        """
+        Return True if any lane has a name.
+
+        Examples::
+        >>> heat = Heat(event="1", heat=1, lanes=[Lane(name="Alice")])
+        >>> heat.has_names()
+        True
+        >>> heat = Heat(event="1", heat=1, lanes=[Lane()])
+        >>> heat.has_names()
+        False
+        """
+        return any(lane.name for lane in self._lanes)
+
+    def merge(
+        self,
+        info_from: "Heat | None" = None,
+        results_from: "Heat | None" = None,
+    ) -> None:
+        """
+        Merge another Heat object into this one.
+
+        Merging "info" will overwrite:
+        - event, heat, description
+        - Lanes: name, team, seed_time, age
+
+        Merging "results" will overwrite:
+        - meet_id, race, time_recorded, time_resolver
+        - Lanes: times, is_dq, is_empty
+
+        :param info_from: Merge the heat information into this one
+        :param results_from: Merge the race results into this one
+        """
+        if info_from is not None:
+            self.event = info_from.event
+            self.heat = info_from.heat
+            self.description = info_from.description
+            for lane_number in range(1, 11):
+                dest = self.lane(lane_number)
+                dest.merge(info_from=info_from.lane(lane_number))
+
+        if results_from is not None:
+            self.meet_id = results_from.meet_id
+            self.race = results_from.race
+            self.time_recorded = results_from.time_recorded
+            for lane_number in range(1, 11):
+                dest = self.lane(lane_number)
+                dest.merge(results_from=results_from.lane(lane_number))
+
+    def __lt__(self, other: "Heat") -> bool:
+        """
+        Compare two Heat objects for sorting.
+
+        Comparison is done by event number, then by heat number. For the cases
+        where the event is a number followed by an optional string, the number
+        is compared first, then the string. For example, '1Z' comes before
+        '10S'. For events that don't match this pattern, string comparison is
+        used.
+        """
+        # Treat None as empty/zero for the purposes of sorting.
+        self_event = self.event or ""
+        self_heat = self.heat or 0
+        other_event = other.event or ""
+        other_heat = other.heat or 0
+
+        # The event number is a string, composed of a number and an optional
+        # letter. Sort by the number first, then by the letter. For example,
+        # '1Z' comes before '10S'
+        pattern = r"^(\d+)([A-Z]*)$"
+        self_match = re.match(pattern, self_event.upper())
+        other_match = re.match(pattern, other_event.upper())
+        if not self_match or not other_match:
+            # Fallback to string comparison
+            return self_event.upper() < other_event.upper()
+
+        self_number = int(self_match.group(1))
+        other_number = int(other_match.group(1))
+        if self_number != other_number:
+            return self_number < other_number
+        self_letter = self_match.group(2)
+        other_letter = other_match.group(2)
+        if self_letter != other_letter:
+            return self_letter < other_letter
+        return self_heat < other_heat
 
 
 def truncate_hundredths(time: Time) -> Time:
